# api-gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Comprehensive API documentation for the Ecommerce Platform. All endpoints are accessed through the API Gateway which acts as a reverse proxy to backend microservices (Identity Service, Product Service, Order Service, etc.). The Gateway automatically forwards cookies, headers, and user context to backend services.",
        "title": "Ecommerce Platform API Gateway",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/v1",
    "paths": {
...
        "/orders": {
            "get": {
                "description": "Lấy danh sách đơn hàng theo user_id hoặc session_id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Lấy danh sách đơn hàng",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "session_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.OrderResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Tạo đơn hàng (hỗ trợ multi-shop). Hỗ trợ cả user_id (header X-User-Id) hoặc session_id cho guest",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Tạo đơn hàng từ giỏ hàng",
                "parameters": [
                    {
                        "description": "Thông tin tạo đơn",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.OrderResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/orders/number/{order_number}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Lấy đơn hàng theo số đơn",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order number",
                        "name": "order_number",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.OrderResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/orders/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Lấy đơn hàng theo ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.OrderResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CreateOrderRequest": {
            "type": "object",
            "properties": {
                "address_id": {
                    "type": "integer",
                    "example": 11
                },
                "payment_method": {
                    "type": "string",
                    "example": "cod"
                }
            }
        },
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Bad request"
                },
                "message": {
                    "type": "string",
                    "example": "Invalid input"
                }
            }
        },
...
        "models.OrderItem": {
            "type": "object",
            "properties": {
                "price": {
                    "type": "number",
                    "example": 15100
                },
                "product_item_id": {
                    "type": "integer",
                    "example": 1001
                },
                "product_name": {
                    "type": "string",
                    "example": "Găng tay chơi game"
                },
                "quantity": {
                    "type": "integer",
                    "example": 2
                },
                "subtotal": {
                    "type": "number",
                    "example": 30200
                }
            }
        },
        "models.OrderResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2026-01-15T12:05:00Z"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.OrderItem"
                    }
                },
                "order_id": {
                    "type": "integer",
                    "example": 5001
                },
                "order_number": {
                    "type": "string",
                    "example": "ORD-20260115-5001"
                },
                "status": {
                    "type": "string",
                    "example": "pending"
                },
                "total_amount": {
                    "type": "number",
                    "example": 30200
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT access token. Example: \"Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "CookieAuth": {
            "description": "HttpOnly cookie containing JWT access token. Automatically set after login/register. Used for authentication in browser-based applications.",
            "type": "apiKey",
            "name": "access_token",
            "in": "cookie"
        }
    }
}